requirements.txt
# Core dependencies
flask==2.3.3
requests==2.31.0
openai==0.28.0
weaviate-client==3.26.7
numpy==1.24.3

# Development dependencies
pytest==7.4.0
pytest-cov==4.1.0
black==23.7.0
flake8==6.0.0

# Docker integration
docker==6.1.3

# Additional utilities
python-dotenv==1.0.0 